requirements.txt
autopep8==1.5.5   
cacheout==0.11.2  
certifi==2020.12.5
chardet==4.0.0    
click==7.1.2      
colorama==0.4.4   
fastapi==0.63.0   
h11==0.12.0       
idna==2.10
loguru==0.5.3
pycodestyle==2.6.0
pydantic==1.7.3
PyJWT==2.0.1
python-dotenv==0.15.0
PyYAML==5.4.1
requests==2.25.1
SQLAlchemy==1.4.3
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
watchgod==0.6
websockets==8.1
win32-setctime==1.0.3
python-dotenv==0.15.0
google-trans-new==1.1.9
googletrans==2.4.0
#uvicorn main:app --reload --env-file .env
#https://adyenpayment-api-dev-5tx5aqryhq-ue.a.run.app/docs#/